Refactor SteamId64Field into seperate component
