+++ projects/save-editor/components/Modals/Fields/index.tsx
@@ -1,0 +1,6 @@
+export { default as SteamId64Field } from "./SteamId64Field";
+
+export interface FieldProps<T> {
+  initialValue: T;
+  onChange: (value: T | undefined) => void;
+}
