+++ projects/save-editor/components/Modals/PlayerDataModal.tsx
@@ -1,8 +1,7 @@
-import { Button, Input, Modal, ModalProps, Row, Spacer, Text, useInput, useModal } from "@nextui-org/react";
-import React, { useImperativeHandle, useMemo, useState } from "react";
-import Player from "../../structures/player";
-import SteamProfileCell from "../DataGrids/Cells/SteamProfileCell";
-import SteamID from "steamid";
+import { Button, Modal, ModalProps, Spacer, Text, useModal } from "@nextui-org/react";
+import React, { useImperativeHandle, useRef } from "react";
+import Player, { IPlayer } from "../../structures/player";
+import { FieldProps, SteamId64Field } from "./Fields";
 
 export interface PlayerDataModalProps extends ModalProps {
   player: Player;
@@ -14,55 +13,17 @@ const PlayerDataModal = React.forwardRef(({ player, ...props }: PlayerDataModalP
     setVisible,
   }));
 
-  const InputSteamId64 = () => {
-    const { value, reset, bindings } = useInput(`${player.steamId64}`);
-    
-    const helper = useMemo((): any => {
-      try {
-        const id = new SteamID(value)
-        if (!id.isValid()) {
-          return {
-            error: "Invalid Steam ID",
-          }
-        } else if (!id.isValidIndividual()) {
-          return {
-            error: "Steam ID does not belong to an individual",
-          }
-        } else {
-          return {
-            validSteamId64: id.getSteamID64(),
-          }
-        }
-      } catch (error) {
-        return {
-          error: "Invalid Steam ID",
-        }
-      }
-    }, [value]);
-
-    return (<>
-      <Row css={{
-        ".nextui-user-info": {
-          "width": "calc(100% - 4 * var(--nextui-space-sm))",
-          "*": {
-            "width": "100%",
-          }
-        }
-      }}>
-        <SteamProfileCell steamId64={helper.validSteamId64} cacheMissing />
-      </Row>
-      <Input
-        {...bindings}
-        clearable
-        bordered
-        fullWidth
-        onClearClick={reset}
-        helperColor={"error"}
-        helperText={helper.error ?? ""}
-        label="Steam ID"
-      />
-      <Spacer y={0.1} />
-    </>);
+  const values = useRef<Partial<IPlayer>>({});
+  const setValue = <T extends keyof IPlayer>(key: T) => {
+    return (value: IPlayer[T] | undefined) => {
+      values.current![key] = value;
+    }
+  }
+  const getFieldProps = <T extends keyof IPlayer>(key: T): FieldProps<IPlayer[T]> => {
+    return {
+      initialValue: player[key],
+      onChange: setValue(key),
+    }
   }
 
   return (
@@ -73,7 +34,8 @@ const PlayerDataModal = React.forwardRef(({ player, ...props }: PlayerDataModalP
         </Text>
       </Modal.Header>
       <Modal.Body>
-        <InputSteamId64 />
+        <SteamId64Field {...getFieldProps("steamId64")} />
+        <Spacer y={0.1} />
       </Modal.Body>
       <Modal.Footer>
         <Button auto flat color="error" onClick={() => setVisible(false)}>
